clear architecture directory refactor and linter rules added
