+++ lib/src/core/util/big_number.dart
