+++ lib/src/presentation/widget/fixed_search_bar.dart
