+++ lib/src/presentation/widget/interactive_aside_bar_texts.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../utils/ui_constants.dart';
+import '../../core/util/ui_constants.dart';
 
 class InteractiveAsideBarTexts extends StatelessWidget {
   final String textContent;
